+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/WasherHandlerImpl.java
@@ -1,10 +1,13 @@
 package com.dwarfeng.fdr.impl.handler;
 
 import com.dwarfeng.fdr.sdk.handler.WasherMaker;
-import com.dwarfeng.fdr.stack.exception.UnsupportedWasherTypeException;
 import com.dwarfeng.fdr.stack.exception.WasherException;
+import com.dwarfeng.fdr.stack.exception.UnsupportedWasherTypeException;
+import com.dwarfeng.fdr.stack.handler.RecordMemoryHandler;
 import com.dwarfeng.fdr.stack.handler.Washer;
 import com.dwarfeng.fdr.stack.handler.WasherHandler;
+import com.dwarfeng.fdr.stack.struct.RecordMemory;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,9 +22,14 @@ public class WasherHandlerImpl implements WasherHandler {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(WasherHandlerImpl.class);
 
+    private final RecordMemoryHandler recordMemoryHandler;
+
     private final List<WasherMaker> washerMakers;
 
-    public WasherHandlerImpl(List<WasherMaker> washerMakers) {
+    private final InternalWasherContext washerContext = new InternalWasherContext();
+
+    public WasherHandlerImpl(List<WasherMaker> washerMakers, RecordMemoryHandler recordMemoryHandler) {
+        this.recordMemoryHandler = recordMemoryHandler;
         this.washerMakers = Optional.ofNullable(washerMakers).orElse(Collections.emptyList());
     }
 
@@ -33,6 +41,7 @@ public class WasherHandlerImpl implements WasherHandler {
             WasherMaker washerMaker = washerMakers.stream().filter(maker -> maker.supportType(type))
                     .findFirst().orElseThrow(() -> new UnsupportedWasherTypeException(type));
             Washer washer = washerMaker.makeWasher(type, param);
+            washer.init(washerContext);
             LOGGER.debug("清洗器构建成功!");
             LOGGER.debug("清洗器: {}", washer);
             return washer;
@@ -40,6 +49,14 @@ public class WasherHandlerImpl implements WasherHandler {
             throw e;
         } catch (Exception e) {
             throw new WasherException(e);
+        }
+    }
+
+    private final class InternalWasherContext implements Washer.Context {
+
+        @Override
+        public List<RecordMemory> lookupRecordMemory(LongIdKey pointKey) throws Exception {
+            return recordMemoryHandler.lookup(pointKey);
         }
     }
 }